Resolve #5 - workaround for canceled keychain.bashrc source.
Instead of trying to ignore/bypass the error on canceling, have the code block in .bashrc source all files EXCEPT keychain.bashrc. Then, source keychain.bashrc.

The caveat of this method, while it mostly resolves this issue, is that all subsequent code under that source block will not be read if keychain.bashrc is canceled. However, due to this problem, I have added a comment within the overall block to let users know where to put other configuration as needed.